requirements.txt
pymongo
requests
python-dotenv
websocket-client
python-kraken-sdk